requirements.txt
pyarrow==22.0.0
apache-airflow[amazon]==3.0.6
apache-airflow-providers-postgres==6.4.1
apache-airflow-providers-slack==9.3.0
pandas>=2.1.0,<2.2.0
numpy>=1.26.0,<2.0.0
requests==2.32.5
pydantic-settings==2.11.0
boto3==1.40.59
black==25.11.0
flake8==7.3.0
pytest==9.0.1
pytest-mock==3.15.1
pytest-cov==7.0.0




